+++ packages/core/src/Aesthetic.tsx
@@ -84,7 +84,7 @@ export default abstract class Aesthetic<
     const globalSheet = globalDef ? globalDef(this.getTheme()) : null;
 
     if (globalSheet) {
-      const sheet = this.processStyleSheet(
+      const sheet = this.parseStyleSheet(
         this.syntax.convertGlobalSheet(globalSheet, options).toObject(),
         ':root',
       );
@@ -109,22 +109,22 @@ export default abstract class Aesthetic<
 
     this.applyGlobalStyles(options);
 
-    const baseSheet = this.syntax.convertStyleSheet(this.getStyleSheet(styleName), {
+    const nativeSheet = this.syntax.convertStyleSheet(this.getStyleSheet(styleName), {
       ...options,
       name: styleName,
     });
-    const styleSheet = this.processStyleSheet(baseSheet.toObject(), styleName);
+    const parsedSheet = this.parseStyleSheet(nativeSheet.toObject(), styleName);
 
     this.cache[styleName] = {
-      ...styleSheet,
-      ...baseSheet.classNames,
+      ...parsedSheet,
+      ...nativeSheet.classNames,
     } as SheetMap<ParsedBlock>;
 
     return this.cache[styleName];
   }
 
   /**
-   * Compose and extend multiple stylesheets to create 1 stylesheet.
+   * Compose and extend multiple style sheets to create 1 style sheet.
    */
   extendStyles(
     ...styleSheets: StyleSheetDefinition<Theme, any>[]
@@ -158,8 +158,8 @@ export default abstract class Aesthetic<
   flushStyles(styleName: StyleName) {}
 
   /**
-   * Retrieve the defined component styles. If the definition is a function,
-   * execute it while passing the current theme and React props.
+   * Retrieve the defined component style sheet. If the definition is a function,
+   * execute it while passing the current theme.
    */
   getStyleSheet(styleName: StyleName): StyleSheet {
     const parentStyleName = this.parents[styleName];
@@ -215,7 +215,15 @@ export default abstract class Aesthetic<
   }
 
   /**
-   * Register a style sheet definition for a unique component by name.
+   * Parse an Aesthetic style sheet into an adapter native style sheet.
+   */
+  parseStyleSheet(styleSheet: SheetMap<NativeBlock>, styleName: StyleName): SheetMap<ParsedBlock> {
+    // @ts-ignore Allow spread
+    return { ...styleSheet };
+  }
+
+  /**
+   * Register a style sheet definition. Optionally extend from a parent style sheet if defined.
    */
   registerStyleSheet<T>(
     styleName: StyleName,
@@ -240,7 +248,8 @@ export default abstract class Aesthetic<
   }
 
   /**
-   * Register a theme with a pre-defined set of theme settings.
+   * Register a theme with a set of parameters. Optionally register
+   * a global style sheet to apply to the entire document.
    */
   registerTheme<T>(
     themeName: ThemeName,
@@ -262,7 +271,7 @@ export default abstract class Aesthetic<
   }
 
   /**
-   * Transform the list of style declarations to a list of class name.
+   * Transform the list of style objects to a list of CSS class names.
    */
   transformStyles = (
     styles: (undefined | false | ClassName | NativeBlock | ParsedBlock)[],
@@ -306,9 +315,7 @@ export default abstract class Aesthetic<
         counter += 1;
       });
 
-      parsedBlocks.push(
-        ...Object.values(this.processStyleSheet(nativeSheet.toObject(), inlineName)),
-      );
+      parsedBlocks.push(...Object.values(this.parseStyleSheet(nativeSheet.toObject(), inlineName)));
     }
 
     // Transform parsed blocks to class names
@@ -325,7 +332,7 @@ export default abstract class Aesthetic<
   };
 
   /**
-   * Transform the parsed styles into CSS class names.
+   * Transform the parsed style objects into CSS class names.
    */
   abstract transformToClassName(styles: ParsedBlock[]): ClassName;
 
@@ -340,17 +347,6 @@ export default abstract class Aesthetic<
     this.sheetManager = new StyleSheetManager();
 
     return this.sheetManager;
-  }
-
-  /**
-   * Process from an Aesthetic style sheet to an adapter native style sheet.
-   */
-  protected processStyleSheet(
-    styleSheet: SheetMap<NativeBlock>,
-    styleName: StyleName,
-  ): SheetMap<ParsedBlock> {
-    // @ts-ignore Allow spread
-    return { ...styleSheet };
   }
 
   /**